moved inspect_string_without_overlap  into separate file
