+++ dlp/snippets/Custom_Infotype/inspect_string_without_overlap_test.py
@@ -14,8 +14,18 @@
 
 import os
 
+import inspect_string_without_overlap as custom_infotype
+
 import pytest
 
-import custom_infotype
-
 GCLOUD_PROJECT = os.getenv("GOOGLE_CLOUD_PROJECT")
+
+
+def test_inspect_string_without_overlap(capsys: pytest.LogCaptureFixture) -> None:
+    custom_infotype.inspect_string_without_overlap(
+        GCLOUD_PROJECT, "example.com is a domain, james@example.org is an email."
+    )
+
+    out, _ = capsys.readouterr()
+    assert "example.com" in out
+    assert "example.org" not in out
